Add sub command deleteMigrations. Add config for repeated paths
