+++ tools/cli/config.js
@@ -1,0 +1,7 @@
+const migrationsDir = '/packages/server/src/database/migrations';
+const seedsDir = '/packages/server/src/database/seeds';
+
+exports.modules = {
+  migrationsDir,
+  seedsDir
+};
